requirements.txt
py-clob-client
python-dotenv
pandas
gspread
gspread-dataframe
sortedcontainers
black==24.4.2
eth-account
eth-utils===4.1.1
poly_eip712_structs==0.0.1
py_order_utils==0.3.2
pytest==8.2.2
python-dotenv==0.19.2
requests==2.32.3
websockets==12.0
cryptography==42.0.8
asyncio
requests
google-auth
web3